feat(conf): enable use of negative numbers in offset setting (#219)

+++ crates/head/src/query.rs
@@ -8,7 +8,7 @@ use std::str::FromStr;
 #[struct_merge("crate::settings::Settings")]
 pub struct SettingsMerge {
     pub scroll_speed: Option<u16>,
-    pub offset: Option<u16>,
+    pub offset: Option<i8>,
     pub judge_position: Option<i32>,
     pub scroll_direction: Option<ScrollDirection>,
     pub lane_gap: Option<u8>,
@@ -51,7 +51,7 @@ pub fn get_optional_settings() -> SettingsMerge {
                         settings.scroll_speed.replace(speed);
                     }
                     "offset" => {
-                        let offset = value.parse::<u16>().unwrap();
+                        let offset = value.parse::<i8>().unwrap();
                         settings.offset.replace(offset);
                     }
                     "scroll_direction" => {
